wrote NightModeService to use in header and footer
