+++ src/app/components/footer/footer.component.html
@@ -3,7 +3,7 @@
         2022 &copy; Course Gamification is licensed under CC BY-NC-SA 4.0
         <a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/4.0/">
             <img alt="Creative Commons License"
-                 [src]="isNightMode()? 'assets/global/cc%20by-nc-sa%20dark%20mode.png' : 'assets/global/cc%20by-nc-sa.png'"
+                 [src]="getNightMode()? 'assets/global/cc%20by-nc-sa%20dark%20mode.png' : 'assets/global/cc%20by-nc-sa.png'"
                  style="float: right; height: 25px"
                  class="px-2 py-1 h-15"
             />
